Separated RL tests into new file
